+++ HISTORY.rst
@@ -8,6 +8,8 @@ Unreleased
 
 **Features and Improvements**
 
+* ALCYN-1772: Add manufacturer information in product export (ESB)
+
 **Bugfixes**
 * Fix additional products and pushed moves. Limit additional products to pick only